feat: create separate CreateMatchContent widget and refactor CreateMatchScreen
